requirements.txt
paho-mqtt==1.5.0
python-dotenv==0.10.3
redis==3.3.11
websocket-client==0.56.0
websockets==8.1
